+++ see/__init__.py
@@ -1,21 +1,17 @@
 """
-see
-dir for humans
+see: dir for humans.
 
-    >>> from see import see
-    >>> help(see)
-
-Copyright (c) 2009-2017 Liam Cooke
-https://araile.github.io/see/
+Documentation is available at https://araile.github.io/see/
 
 """
 from .inspector import see
+from .output import SeeResult
 
 
-__all__ = ['see']
+__all__ = ['see', 'SeeResult']
 
 __author__ = 'Liam Cooke'
 __contributors__ = 'See AUTHORS.rst'
-__version__ = '1.4.0a0'
+__version__ = '1.4.0'
 __copyright__ = 'Copyright (c) 2009-2017 Liam Cooke'
 __license__ = 'BSD License'