+++ see/tools.py
@@ -1,13 +1,9 @@
 """
-see.tools
-Filtering and other tasks
-
-Copyright (c) 2009-2017 Liam Cooke
-https://araile.github.io/see/
+Filtering and other tasks.
 
 """
 import fnmatch
-import re
+import unicodedata
 
 
 def compact(objects):
@@ -17,23 +13,40 @@ def compact(objects):
     return tuple(filter(bool, objects or []))
 
 
-def filter_regex(names, pat):
+def char_width(char):
+    """
+    Get the display length of a unicode character.
+    """
+    if ord(char) < 128:
+        return 1
+    elif unicodedata.east_asian_width(char) in ('F', 'W'):
+        return 2
+    elif unicodedata.category(char) in ('Mn',):
+        return 0
+    else:
+        return 1
+
+
+def display_len(text):
+    """
+    Get the display length of a string. This can differ from the character
+    length if the string contains wide characters.
+    """
+    text = unicodedata.normalize('NFD', text)
+    return sum(char_width(char) for char in text)
+
+
+def filter_regex(names, regex):
     """
     Return a tuple of strings that match the regular expression pattern.
     """
-    pat = re.compile(pat)
-
-    def match(name, fn=pat.search):
-        return fn(name) is not None
-
-    return tuple(filter(match, names))
+    return tuple(name for name in names
+                 if regex.search(name) is not None)
 
 
-def filter_wildcard(names, pat):
+def filter_wildcard(names, pattern):
     """
-    Return a tuple of strings that match a shell-style pattern.
+    Return a tuple of strings that match a shell-style wildcard pattern.
     """
-    def match(name, fn=fnmatch.fnmatch, pat=pat):
-        return fn(name, pat)
-
-    return tuple(filter(match, names))
+    return tuple(name for name in names
+                 if fnmatch.fnmatch(name, pattern))